refactor: to use generic type
