--- src/main/java/com/yourssu/roomescape/constants/RoleConstants.java
@@ -1,6 +1,0 @@
-package com.yourssu.roomescape.constants;
-
-public class RoleConstants {
-    public static final String ADMIN = "ADMIN";
-    public static final String USER = "USER";
-}
