+++ src/main/java/com/yourssu/roomescape/enums/Role.java
@@ -1,0 +1,16 @@
+package com.yourssu.roomescape.constants;
+
+public enum Role {
+    ADMIN("ADMIN"),
+    USER("USER");
+
+    private final String value;
+
+    Role(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
+}
