+++ src/main/java/com/yourssu/roomescape/reservation/ReservationController.java
@@ -24,7 +24,7 @@ public class ReservationController {
     }
 
     @PostMapping("/reservations")
-    public ResponseEntity create(@RequestBody ReservationRequest reservationRequest, @LoginMember Member member) {
+    public ResponseEntity<ReservationResponse> create(@RequestBody ReservationRequest reservationRequest, @LoginMember Member member) {
         if (reservationRequest.getDate() == null
             || reservationRequest.getTheme() == null
             || reservationRequest.getTime() == null) {
@@ -36,7 +36,7 @@ public class ReservationController {
     }
 
     @DeleteMapping("/reservations/{id}")
-    public ResponseEntity delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Long id) {
         reservationService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
